+++ src/components/pool/PoolWarningTooltip.vue
@@ -4,7 +4,6 @@ import { PoolWarning, usePoolWarning } from '@/composables/usePoolWarning';
 import { Pool } from '@/services/pool/types';
 import { computed } from 'vue';
 import { useI18n } from 'vue-i18n';
-import { isSoftMigratablePool } from '../forms/pool_actions/MigrateForm/constants';
 
 /**
  * TYPES
@@ -29,7 +28,7 @@ const { isAffectedBy } = usePoolWarning(poolId);
  * COMPUTED
  */
 const warningLabel = computed((): string => {
-  if (isMigratablePool(props.pool) && !isSoftMigratablePool(props.pool.id)) {
+  if (isMigratablePool(props.pool)) {
     return t('deprecatedPool.tooltip');
   }
